+++ aimiaobi-20230801/alibabacloud_aimiaobi20230801/models/_create_dataset_request.py
@@ -10,6 +10,7 @@ from darabonba.model import DaraModel
 class CreateDatasetRequest(DaraModel):
     def __init__(
         self,
+        access_level: str = None,
         dataset_config: main_models.CreateDatasetRequestDatasetConfig = None,
         dataset_description: str = None,
         dataset_name: str = None,
@@ -19,6 +20,7 @@ class CreateDatasetRequest(DaraModel):
         search_dataset_enable: int = None,
         workspace_id: str = None,
     ):
+        self.access_level = access_level
         self.dataset_config = dataset_config
         self.dataset_description = dataset_description
         # This parameter is required.
@@ -41,6 +43,9 @@ class CreateDatasetRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.access_level is not None:
+            result['AccessLevel'] = self.access_level
+
         if self.dataset_config is not None:
             result['DatasetConfig'] = self.dataset_config.to_map()
 
@@ -69,6 +74,9 @@ class CreateDatasetRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('AccessLevel') is not None:
+            self.access_level = m.get('AccessLevel')
+
         if m.get('DatasetConfig') is not None:
             temp_model = main_models.CreateDatasetRequestDatasetConfig()
             self.dataset_config = temp_model.from_map(m.get('DatasetConfig'))