+++ aimiaobi-20230801/alibabacloud_aimiaobi20230801/models/_list_data_permissions_response_body.py
@@ -1,0 +1,191 @@
+# -*- coding: utf-8 -*-
+# This file is auto-generated, don't edit it. Thanks.
+from __future__ import annotations
+
+from typing import List
+
+from alibabacloud_aimiaobi20230801 import models as main_models
+from darabonba.model import DaraModel
+
+class ListDataPermissionsResponseBody(DaraModel):
+    def __init__(
+        self,
+        code: str = None,
+        data: List[main_models.ListDataPermissionsResponseBodyData] = None,
+        http_status_code: int = None,
+        message: str = None,
+        page_number: int = None,
+        page_size: int = None,
+        request_id: str = None,
+        success: bool = None,
+        total_count: int = None,
+    ):
+        self.code = code
+        self.data = data
+        self.http_status_code = http_status_code
+        self.message = message
+        self.page_number = page_number
+        self.page_size = page_size
+        self.request_id = request_id
+        self.success = success
+        self.total_count = total_count
+
+    def validate(self):
+        if self.data:
+            for v1 in self.data:
+                 if v1:
+                    v1.validate()
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.code is not None:
+            result['Code'] = self.code
+
+        result['Data'] = []
+        if self.data is not None:
+            for k1 in self.data:
+                result['Data'].append(k1.to_map() if k1 else None)
+
+        if self.http_status_code is not None:
+            result['HttpStatusCode'] = self.http_status_code
+
+        if self.message is not None:
+            result['Message'] = self.message
+
+        if self.page_number is not None:
+            result['PageNumber'] = self.page_number
+
+        if self.page_size is not None:
+            result['PageSize'] = self.page_size
+
+        if self.request_id is not None:
+            result['RequestId'] = self.request_id
+
+        if self.success is not None:
+            result['Success'] = self.success
+
+        if self.total_count is not None:
+            result['TotalCount'] = self.total_count
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('Code') is not None:
+            self.code = m.get('Code')
+
+        self.data = []
+        if m.get('Data') is not None:
+            for k1 in m.get('Data'):
+                temp_model = main_models.ListDataPermissionsResponseBodyData()
+                self.data.append(temp_model.from_map(k1))
+
+        if m.get('HttpStatusCode') is not None:
+            self.http_status_code = m.get('HttpStatusCode')
+
+        if m.get('Message') is not None:
+            self.message = m.get('Message')
+
+        if m.get('PageNumber') is not None:
+            self.page_number = m.get('PageNumber')
+
+        if m.get('PageSize') is not None:
+            self.page_size = m.get('PageSize')
+
+        if m.get('RequestId') is not None:
+            self.request_id = m.get('RequestId')
+
+        if m.get('Success') is not None:
+            self.success = m.get('Success')
+
+        if m.get('TotalCount') is not None:
+            self.total_count = m.get('TotalCount')
+
+        return self
+
+class ListDataPermissionsResponseBodyData(DaraModel):
+    def __init__(
+        self,
+        create_time: str = None,
+        create_user: str = None,
+        data_id: str = None,
+        data_type: str = None,
+        id: int = None,
+        permission: str = None,
+        user_id: str = None,
+        username: str = None,
+    ):
+        self.create_time = create_time
+        self.create_user = create_user
+        self.data_id = data_id
+        self.data_type = data_type
+        self.id = id
+        self.permission = permission
+        self.user_id = user_id
+        self.username = username
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        result = dict()
+        _map = super().to_map()
+        if _map is not None:
+            result = _map
+        if self.create_time is not None:
+            result['CreateTime'] = self.create_time
+
+        if self.create_user is not None:
+            result['CreateUser'] = self.create_user
+
+        if self.data_id is not None:
+            result['DataId'] = self.data_id
+
+        if self.data_type is not None:
+            result['DataType'] = self.data_type
+
+        if self.id is not None:
+            result['Id'] = self.id
+
+        if self.permission is not None:
+            result['Permission'] = self.permission
+
+        if self.user_id is not None:
+            result['UserId'] = self.user_id
+
+        if self.username is not None:
+            result['Username'] = self.username
+
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('CreateTime') is not None:
+            self.create_time = m.get('CreateTime')
+
+        if m.get('CreateUser') is not None:
+            self.create_user = m.get('CreateUser')
+
+        if m.get('DataId') is not None:
+            self.data_id = m.get('DataId')
+
+        if m.get('DataType') is not None:
+            self.data_type = m.get('DataType')
+
+        if m.get('Id') is not None:
+            self.id = m.get('Id')
+
+        if m.get('Permission') is not None:
+            self.permission = m.get('Permission')
+
+        if m.get('UserId') is not None:
+            self.user_id = m.get('UserId')
+
+        if m.get('Username') is not None:
+            self.username = m.get('Username')
+
+        return self
+