+++ aimiaobi-20230801/alibabacloud_aimiaobi20230801/models/_update_dataset_shrink_request.py
@@ -7,12 +7,14 @@ from darabonba.model import DaraModel
 class UpdateDatasetShrinkRequest(DaraModel):
     def __init__(
         self,
+        access_level: str = None,
         dataset_config_shrink: str = None,
         dataset_description: str = None,
         dataset_id: int = None,
         search_dataset_enable: int = None,
         workspace_id: str = None,
     ):
+        self.access_level = access_level
         self.dataset_config_shrink = dataset_config_shrink
         self.dataset_description = dataset_description
         self.dataset_id = dataset_id
@@ -28,6 +30,9 @@ class UpdateDatasetShrinkRequest(DaraModel):
         _map = super().to_map()
         if _map is not None:
             result = _map
+        if self.access_level is not None:
+            result['AccessLevel'] = self.access_level
+
         if self.dataset_config_shrink is not None:
             result['DatasetConfig'] = self.dataset_config_shrink
 
@@ -47,6 +52,9 @@ class UpdateDatasetShrinkRequest(DaraModel):
 
     def from_map(self, m: dict = None):
         m = m or dict()
+        if m.get('AccessLevel') is not None:
+            self.access_level = m.get('AccessLevel')
+
         if m.get('DatasetConfig') is not None:
             self.dataset_config_shrink = m.get('DatasetConfig')